% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ggsurvplot.R
\name{+.ggsurvplot}
\alias{+.ggsurvplot}
\alias{\%+\%}
\title{Add Components to a ggsurvplot}
\usage{
\method{+}{ggsurvplot}(e1, e2)

e1 \%+\% e2
}
\arguments{
\item{e1}{an object of class \code{\link{ggsurvplot}}.}

\item{e2}{a theme function.}
}
\description{
Allows to add ggplot theme function to an object of class ggsurvplot.
}
\examples{
# Fit survival curves
require("survival")
fit<- survfit(Surv(time, status) ~ sex, data = lung)

# Basic survival curves
p <- ggsurvplot(fit, data = lung, risk.table = TRUE,
   main = "Survival curve",
   submain = "Based on Kaplan-Meier estimates",
   caption = "created with survminer"
   )
p

# Customizing the plots
p \%+\% theme_survminer(
     font.main = c(16, "bold", "darkblue"),
     font.submain = c(15, "bold.italic", "purple"),
     font.caption = c(14, "plain", "orange"),
     font.x = c(14, "bold.italic", "red"),
     font.y = c(14, "bold.italic", "darkred"),
     font.tickslab = c(12, "plain", "darkgreen")
)

}
