% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_classic2.R
\name{theme_classic2}
\alias{theme_classic2}
\title{ggplot2 classic theme with axis lines}
\usage{
theme_classic2(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Create a ggplot2 classic theme with axis lines.
}
\examples{
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
   geom_point(aes(color = gear))

# Default plot
p

# Use theme_classic()
p + theme_classic()

# Use theme_classic2()
p + theme_classic2()
}
