% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_survminer.R
\name{theme_survminer}
\alias{theme_survminer}
\title{Default Theme for Survminer Plots}
\usage{
theme_survminer(base_size = 12, base_family = "", font.main = c(16,
  "plain", "black"), font.submain = c(15, "plain", "black"), font.x = c(14,
  "plain", "black"), font.y = c(14, "plain", "black"), font.caption = c(15,
  "plain", "black"), font.tickslab = c(12, "plain", "black"),
  font.legend = c(10, "plain", "black"))
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{font.main, font.submain, font.caption, font.x, font.y, font.tickslab, font.legend}{a vector of length 3
indicating respectively the size (e.g.: 14), the style (e.g.: "plain",
"bold", "italic", "bold.italic") and the color (e.g.: "red") of main title, subtitle, caption,
xlab and ylab, axis tick labels and legend, respectively. For example \emph{font.x =
c(14, "bold", "red")}.  Use font.x = 14, to change only font size; or use
font.x = "bold", to change only font face.}
}
\description{
Default theme for plots generated with survminer.
}
\examples{

# Fit survival curves
#++++++++++++++++++++++++++++++++++++
require("survival")
fit<- survfit(Surv(time, status) ~ sex, data = lung)

# Basic survival curves
#++++++++++++++++++++++++++++++++++++
ggsurvplot(fit, data = lung)

# Change font size, style and color
#++++++++++++++++++++++++++++++++++++
# Change font size, style and color at the same time
# Use font.x = 14, to change only font size; or use
# font.x = "bold", to change only font face.

ggsurvplot(
   fit, data = lung,
   main = "Survival curve",
   submain = "Based on Kaplan-Meier estimates",
   caption = "created with survminer",

   ggtheme = theme_survminer(
     font.main = c(16, "bold", "darkblue"),
     font.submain = c(15, "bold.italic", "purple"),
     font.caption = c(14, "plain", "orange"),
     font.x = c(14, "bold.italic", "red"),
     font.y = c(14, "bold.italic", "darkred"),
     font.tickslab = c(12, "plain", "darkgreen")
   )
 )
}
