% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggadjustedcurves.R
\name{ggadjustedcurves}
\alias{ggadjustedcurves}
\title{Adjusted Survival Curves for Cox Proportional Hazards Model}
\usage{
ggadjustedcurves(fit, variable = NULL, data = NULL, reference = NULL,
  method = "conditional", fun = NULL, palette = "hue",
  ylab = "Survival rate", ggtheme = theme_survminer(), ...)
}
\arguments{
\item{fit}{an object of class \link{coxph.object} - created with \link{coxph} function.}

\item{variable}{a character, name of the grouping variable to be plotted. If not suplied then it will be extracted from the model formula from the \code{strata()} component. If there is no \code{strata()} component then only a single curve will be plotted - average for the population.}

\item{data}{a dataset for predictions. If not supplied then data will be extracted from the \code{fit} object.}

\item{reference}{a dataset for with reference population, to which dependent variables should be balanced. If not specified, then the \code{data} will be used instead. Note that the \code{reference} dataset should contain variables used in \code{fit} object.}

\item{fun}{an arbitrary function defining a transformation of the survival
curve.  Often used transformations can be specified with a character
argument: "event" plots cumulative events (f(y) = 1-y), "cumhaz" plots the
cumulative hazard function (f(y) = -log(y)), and "pct" for survival
probability in percentage.}

\item{palette}{the color palette to be used. Allowed values include "hue" for
the default hue color scale; "grey" for grey color palettes; brewer palettes
e.g. "RdBu", "Blues", ...; or custom color palette e.g. c("blue", "red").
See details section for more information.  Can be also a numeric vector of
length(groups); in this case a basic color palette is created using the
function \link[grDevices]{palette}.}

\item{ylab}{a label for oy axis.}

\item{ggtheme}{function, ggplot2 theme name. Allowed values include ggplot2 official themes: see \code{\link[ggplot2]{theme}}.}

\item{...}{further arguments passed to the function \code{\link[ggpubr]{ggpar}} for customizing the plot.}
}
\value{
Returns an object of class \code{gg}.
}
\description{
This function plots adjusted survival curves for the \code{coxph} model.
The idea behind this function is described in \code{https://cran.r-project.org/web/packages/survival/vignettes/adjcurve.pdf}.
For every observation in the dataset a prediction for survival curve is made.
Then the predictions are averaged with respect to a selected variable.
}
\examples{

library(survival)
fit2 <- coxph( Surv(stop, event) ~  size, data = bladder )
# single curve
ggadjustedcurves(fit2, data = bladder)

fit2 <- coxph( Surv(stop, event) ~  size + strata(rx), data = bladder )
# average in groups
ggadjustedcurves(fit2, data = bladder, method = "average", variable = "rx")

# conditional balancing in groups
ggadjustedcurves(fit2, data = bladder, method = "conditional", variable = "rx")

# marginal balancing in groups
ggadjustedcurves(fit2, data = bladder, method = "marginal", variable = "rx")

\dontrun{
# this will take some time
fdata <- flchain[flchain$futime >=7,]
fdata$age2 <- cut(fdata$age, c(0,54, 59,64, 69,74,79, 89, 110),
                  labels = c(paste(c(50,55,60,65,70,75,80),
                                   c(54,59,64,69,74,79,89), sep='-'), "90+"))
fdata$group <- factor(1+ 1*(fdata$flc.grp >7) + 1*(fdata$flc.grp >9),
                      levels=1:3,
                      labels=c("FLC < 3.38", "3.38 - 4.71", "FLC > 4.71"))
# single curve
fit <- coxph( Surv(futime, death) ~ age*sex, data = fdata)
ggadjustedcurves(fit, data = fdata, method = "single")

# average in groups
fit <- coxph( Surv(futime, death) ~ age*sex + strata(group), data = fdata)
ggadjustedcurves(fit, data = fdata, method = "average")

# conditional balancing in groups
ggadjustedcurves(fit, data = fdata, method = "conditional", reference = fdata)

# marginal balancing in groups
ggadjustedcurves(fit, data = fdata, method = "marginal")
}

}
\author{
Przemyslaw Biecek, \email{przemyslaw.biecek@gmail.com}
}
