% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcoxzph.R
\name{ggcoxzph}
\alias{ggcoxzph}
\alias{print.ggcoxzph}
\title{Graphical Test of Proportional Hazards using ggplot2}
\usage{
ggcoxzph(fit, resid = TRUE, se = TRUE, df = 4, nsmo = 40, var,
  ggtheme = ggplot2::theme_classic())

\method{print}{ggcoxzph}(x, ...)
}
\arguments{
\item{fit}{an object of class \link{cox.zph}.}

\item{resid}{a logical value, if TRUE the residuals are included on the plot, as well as the smooth fit.}

\item{se}{a logical value, if TRUE, confidence bands at two standard errors will be added.}

\item{df}{the degrees of freedom for the fitted natural spline, df=2 leads to a linear fit.}

\item{nsmo}{number of points used to plot the fitted spline.}

\item{var}{the set of variables for which plots are desired. By default, plots are produced in turn for each variable of a model.}

\item{ggtheme}{function, ggplot2 theme name. Default value is theme_classic().
Allowed values include ggplot2 official themes: theme_gray(), theme_bw(),
theme_minimal(), theme_classic(), theme_void(), ....}

\item{x}{an object of class ggcoxzph}

\item{...}{further arguments passed to print, but really it's unused}
}
\value{
Returns an object of class \code{ggcoxzph} which is a list of ggplots.
}
\description{
Displays a graph of the scaled Schoenfeld residuals, along with a smooth curve using ggplot2. Wrapper around \link{plot.cox.zph}.
}
\section{Functions}{
\itemize{
\item \code{ggcoxzph}: Graphical Test of Proportional Hazards using ggplot2.
}}
\examples{

library(survival)
fit <- coxph(Surv(futime, fustat) ~ age + ecog.ps, data=ovarian)
cox.zph.fit <- cox.zph(fit)
ggcoxzph(cox.zph.fit)

}
\author{
Marcin Kosinski , \email{m.p.kosinski@gmail.com}
}

