% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcoxzph.R
\name{ggcoxzph}
\alias{ggcoxzph}
\alias{print.ggcoxzph}
\title{Graphical Test of Proportional Hazards using ggplot2}
\usage{
ggcoxzph(fit, resid = TRUE, se = TRUE, df = 4, nsmo = 40, var,
  point.col = "red", point.size = 1, point.shape = 19, point.alpha = 1,
  font.main = c(16, "plain", "black"), font.x = c(14, "plain", "black"),
  font.y = c(14, "plain", "black"), font.tickslab = c(12, "plain", "black"),
  ggtheme = theme_classic2())

\method{print}{ggcoxzph}(x, ...)
}
\arguments{
\item{fit}{an object of class \link{cox.zph}.}

\item{resid}{a logical value, if TRUE the residuals are included on the plot,
as well as the smooth fit.}

\item{se}{a logical value, if TRUE, confidence bands at two standard errors
will be added.}

\item{df}{the degrees of freedom for the fitted natural spline, df=2 leads to
a linear fit.}

\item{nsmo}{number of points used to plot the fitted spline.}

\item{var}{the set of variables for which plots are desired. By default, plots
are produced in turn for each variable of a model.}

\item{point.col, point.size, point.shape, point.alpha}{color, size, shape and visibility to be used for points.}

\item{font.main, font.x, font.y, font.tickslab}{a vector of length 3
indicating respectively the size (e.g.: 14), the style (e.g.: "plain",
"bold", "italic", "bold.italic") and the color (e.g.: "red") of main title,
xlab and ylab and axis tick labels, respectively. For example \emph{font.x =
c(14, "bold", "red")}.  Use font.x = 14, to change only font size; or use
font.x = "bold", to change only font face.}

\item{ggtheme}{function, ggplot2 theme name. Default value is survminer::theme_classic2().
Allowed values include ggplot2 official themes: theme_gray(), theme_bw(),
theme_minimal(), theme_classic(), theme_void(), ....}

\item{x}{an object of class ggcoxzph}

\item{...}{further arguments passed to print, but really it's unused}
}
\value{
Returns an object of class \code{ggcoxzph} which is a list of ggplots.
}
\description{
Displays a graph of the scaled Schoenfeld residuals, along with a
 smooth curve using ggplot2. Wrapper around \link{plot.cox.zph}.
}
\section{Functions}{
\itemize{
\item \code{ggcoxzph}: Graphical Test of Proportional Hazards using ggplot2.
}}
\examples{

library(survival)
fit <- coxph(Surv(futime, fustat) ~ age + ecog.ps, data=ovarian)
cox.zph.fit <- cox.zph(fit)
ggcoxzph(cox.zph.fit)

}
\author{
Marcin Kosinski , \email{m.p.kosinski@gmail.com}
}

