% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggforest.R
\name{ggforest}
\alias{ggforest}
\title{Drawing Forest Plot for CoxPH model}
\usage{
ggforest(model, alpha = 0.05)
}
\arguments{
\item{alpha}{significance level.}

\item{fit}{an object of class coxph.}
}
\value{
return an object of class ggplot
}
\description{
Drawing Forest Plot for CoxPH model
}
\examples{
require("survival")
model <- coxph( Surv(time, status) ~ sex + rx + adhere,
                data = colon )
ggforest(model)

}

